feat: Refactor getUser API endpoint and add updateUserName and updateUserPW endpoints
